agents-service/functions/initialize_db/requirements.txt
psycopg2-binary==2.9.9
Faker==19.3.0